refactoring JAXWS components processing plus renaming classes
